# Conflicts:
#	TombEngine/Renderer/Renderer11.cpp
#	TombEngine/Renderer/Renderer11Draw.cpp
#	TombEngine/TombEngine.vcxproj.user

+++ TombEngine/Objects/TR1/Entity/tr1_ape.cpp
@@ -169,7 +169,7 @@ namespace TEN::Entities::Creatures::TR1
 		if (item->HitPoints <= 0)
 		{
 			if (item->Animation.ActiveState != APE_STATE_DEATH)
-				SetAnimation(item, ApeDeathAnims[Random::GenerateInt(0, ApeDeathAnims.size() - 1)]);
+				SetAnimation(item, ApeDeathAnims[Random::GenerateInt(0, (int)ApeDeathAnims.size() - 1)]);
 		}
 		else
 		{
